Render the NewsList title in NewsList
Removed NewsListTitle. Easier to maintain.

Also simplified the props in NewsListStore. Now empty "publisher_code"
means to show all news.
