+++ src/app/routes/PublisherNewsListPage.jsx
@@ -5,7 +5,6 @@ import React from 'react';
 import RouteActions from '../actions/RouteActions';
 import PageMaster from '../routes/PageMaster';
 import NewsList from '../components/NewsList';
-import NewsListTitle from '../components/NewsListTitle';
 
 class PublisherNewsListPage extends PageMaster {
   renderContent() {
@@ -13,7 +12,6 @@ class PublisherNewsListPage extends PageMaster {
     RouteActions.pagePublisherNews(code);
     return (
       <div>
-        <NewsListTitle />
         <NewsList />
       </div>
     );
